refactor(governance): 拆分第三批历史近阈值前端热路径
